server/requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email-validator]
python-dotenv 
psycopg2-binary 
qrcode[pil]
"psycopg[binary]"
sqlalchemy[asyncio]
asyncpg
pydantic[email]
python-jose[cryptography]
pydantic-settings
